gui: relocate whole-GUI functions out of common/, and keep the _gui suffix on mirrors
RELOCATE step: a function whose entire body is a dialog belongs with the dialogs.

  common/styles.c  dt_styles_create_from_list()  ->  gui/common/styles_gui.c

It opens dt_gui_styles_dialog_new() per image and reports "no image selected!" -- no
database work at all. Its only caller was already libs/styles.c. common/styles.c drops
#include "gui/styles.h" with it and now has zero GUI references.

styles is the second real backend/frontend pair, after history_merge.

THE _gui SUFFIX IS LOAD-BEARING, and removing it was my mistake in the previous commit.

Naming the frontend gui/common/styles.h to mirror common/styles.h looks tidier and breaks
the build in a way that is easy to misread. A quoted #include is searched relative to the
INCLUDING FILE's own directory first, so for any file under src/gui/, the spelling

    #include "common/styles.h"

resolves to src/gui/common/styles.h -- not src/common/styles.h. The backend header is
shadowed, and the error surfaces as a pile of implicit-declaration warnings in an unrelated
file (gui/styles_dialog.c could no longer see dt_styles_exists, dt_styles_update, ...),
naming symbols nothing in this commit touched.

Measured before deciding: 47 files under src/gui/ include a common/ header, and the two
mirrors already shadowed 5 of them. The hazard grows with every pair added, which makes
this a property of the convention rather than an accident of these two files.

So mirror files keep a distinguishing suffix: gui/common/<name>_gui.{c,h}. The directory
still says which backend it pairs with, and no filename can shadow its own backend. That
is what the original history_merge_gui.h was doing; I removed the suffix last commit
without understanding why it was there.

Invariant now checked explicitly: no basename appears in both src/common/ and
src/gui/common/. Zero collisions.

Verified: features and nofeatures builds clean, ansel --version runs.
