+++ src/CMakeLists.txt
@@ -84,7 +84,8 @@ FILE(GLOB SOURCE_FILES
   "common/history.c"
   "common/history_actions.c"
   "common/history_merge.c"
-  "gui/common/history_merge.c"
+  "gui/common/history_merge_gui.c"
+  "gui/common/styles_gui.c"
   "common/history_snapshot.c"
   "common/topological_sort.c"
   "common/gpx.c"
