+++ src/gui/actions/styles.c
@@ -26,7 +26,7 @@
 #include "common/act_on.h"
 #include "common/history.h"
 #include "common/history_merge.h"
-#include "gui/common/history_merge.h"
+#include "gui/common/history_merge_gui.h"
 #include "common/styles.h"
 #include "common/usermanual_url.h"
 #include "develop/dev_history.h"
