+++ src/gui/common/history_merge_gui.h
