+++ src/gui/common/styles_gui.h
@@ -1,0 +1,33 @@
+/*
+ *    This file is part of Ansel,
+ *    Copyright (C) 2026 Aurélien PIERRE.
+ *
+ *    Ansel is free software: you can redistribute it and/or modify
+ *    it under the terms of the GNU General Public License as published by
+ *    the Free Software Foundation, either version 3 of the License, or
+ *    (at your option) any later version.
+ *
+ *    Ansel is distributed in the hope that it will be useful,
+ *    but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *    GNU General Public License for more details.
+ *
+ *    You should have received a copy of the GNU General Public License
+ *    along with Ansel.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef DT_GUI_COMMON_STYLES_GUI_H
+#define DT_GUI_COMMON_STYLES_GUI_H
+
+#include <glib.h>
+
+/* The GUI half of common/styles.c: style operations that are a dialog rather than a
+ * database operation. The backend half creates, applies and deletes styles and knows
+ * nothing about windows. */
+
+/** Open the "create style" dialog for each image in `list`, or report that none is
+ *  selected. Purely a dialog driver -- the style itself is created by the backend once
+ *  the user confirms. */
+void dt_styles_create_from_list(const GList *list);
+
+#endif // DT_GUI_COMMON_STYLES_GUI_H
